Refactor telemetry processing classes for improved organization and functionality
- Updated import paths in sentry_options.dart to reflect new structure.
- Introduced new classes for telemetry buffering and processing, including TelemetryBuffer, InMemoryTelemetryBuffer, and DefaultTelemetryProcessor, enhancing the management of telemetry data.
- Added TelemetryBufferConfig for configurable buffer settings.
- Streamlined test imports to align with the new class structure, ensuring consistency across the codebase.
