+++ packages/dart/lib/src/telemetry/processing/buffer.dart
