+++ packages/dart/lib/src/telemetry/processing/buffer_config.dart
