+++ packages/dart/test/telemetry/processing/telemetry_processor_integration_test.dart
@@ -1,7 +1,7 @@
 import 'package:sentry/sentry.dart';
-import 'package:sentry/src/telemetry/processing/in_memory_telemetry_buffer.dart';
-import 'package:sentry/src/telemetry/processing/telemetry_processor.dart';
-import 'package:sentry/src/telemetry/processing/telemetry_processor_integration.dart';
+import 'package:sentry/src/telemetry/processing/in_memory_buffer.dart';
+import 'package:sentry/src/telemetry/processing/processor.dart';
+import 'package:sentry/src/telemetry/processing/processor_integration.dart';
 import 'package:sentry/src/telemetry/span/sentry_span_v2.dart';
 import 'package:test/test.dart';
