+++ packages/dart/test/telemetry/processing/telemetry_processor_test.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:sentry/sentry.dart';
-import 'package:sentry/src/telemetry/processing/telemetry_processor.dart';
+import 'package:sentry/src/telemetry/processing/processor.dart';
 import 'package:sentry/src/telemetry/span/sentry_span_v2.dart';
 import 'package:test/test.dart';
